stringToUtf16List
