+++ json/tokenizer/test.f.cjs
@@ -2,9 +2,10 @@ const tokenizer = require('./module.f.cjs')
 const list = require('../../types/list/module.f.cjs')
 const json = require('../module.f.cjs')
 const { sort } = require('../../types/object/module.f.cjs')
+const encoding = require('../../text/encoding/module.f.cjs');
 
 /** @type {(s: string) => readonly tokenizer.JsonToken[]} */
-const tokenizeString = s => list.toArray(tokenizer.tokenize(list.toCharCodeList(s)))
+const tokenizeString = s => list.toArray(tokenizer.tokenize(encoding.stringToUtf16List(s)))
 
 const stringify = json.stringify(sort)
